+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Permissions page (admin): search, grant, edit and revoke
 
+### Changed
+
+- Email links should always use body in plain text format
+
 ## [1.3.0] - 2025-04-09
 
 ### Added